remove unneeded class
